refactor(memory): update memory system architecture and configuration
- Replace MemorySystem with MemoryStore for improved memory management
- Add MEMORY_SERVICE singleton for centralized memory store access
- Change importance score type from float to int with proper range validation
- Add memory store configuration options (root path, buffer size, cache size)
- Update Note model to use uuid instead of id for memory identification
- Modify test cases to use integer importance values instead of float
- Update Rust backend to use u64 for importance score constants

+++ packages/fabricatio-memory/python/fabricatio_memory/rust.pyi
@@ -5,8 +5,8 @@ import os
 import pathlib
 import typing
 
-MAX_IMPORTANCE_SCORE: builtins.float
-MIN_IMPORTANCE_SCORE: builtins.float
+MAX_IMPORTANCE_SCORE: builtins.int
+MIN_IMPORTANCE_SCORE: builtins.int
 
 @typing.final
 class Memory:
